+++ src/main/java/cart/dto/request/ProductRequest.java
@@ -2,13 +2,13 @@ package cart.dto.request;
 
 public class ProductRequest {
     private String name;
-    private int price;
+    private Long price;
     private String imageUrl;
 
     public ProductRequest() {
     }
 
-    public ProductRequest(String name, int price, String imageUrl) {
+    public ProductRequest(String name, Long price, String imageUrl) {
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
@@ -18,7 +18,7 @@ public class ProductRequest {
         return name;
     }
 
-    public int getPrice() {
+    public Long getPrice() {
         return price;
     }
